requirements.txt
pandas 
python-chess 
requests 
pyarrow
scikit-learn 
matplotlib 
seaborn 
numpy